add tx event helper in api
